backend/requirements.txt
fastapi
uvicorn
easyocr
opencv-python
pillow
transformers
torch
python-multipart
numpy
psycopg2-binary
PyMuPDF
pandas
einops
accelerate
sentencepiece